more comprehensive tests

+++ Sources/BrainflipCLI/IO Helpers/StandardInput.swift
@@ -47,6 +47,6 @@ struct StandardInput: Sequence, IteratorProtocol {
       return nil
     }
 
-    return Unicode.Scalar(UInt32(nextCharacter))
+    return Unicode.Scalar(nextCharacter)
   }
 }
